move comparison to its own file
